+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ Note that ``RB_ID=#`` and ``PHAB_ID=#`` correspond to associated message in comm
 Unreleased
 ----------
 
+* http-server: Add versions of `HttpRouter#filter` which accept a Guice `TypeLiteral` to
+  aid Java users in being able to apply generically typed Filters obtained from the object graph.
+  ``PHAB_ID=D768777``
+
 21.9.0
 ------